Refactored lib/auth/services/firebase_auth_provider.dart -> lib/auth/services/auth_services/firebase_auth_provider.dart
