+++ lib/auth/services/auth_services/auth_service.dart
@@ -1,6 +1,6 @@
 import 'package:echno_attendance/auth/services/auth_services/auth_provider.dart';
 import 'package:echno_attendance/auth/services/auth_services/auth_user.dart';
-import 'package:echno_attendance/auth/services/firebase_auth_provider.dart';
+import 'package:echno_attendance/auth/services/auth_services/firebase_auth_provider.dart';
 
 class AuthService implements AuthProvider {
   final AuthProvider _provider;
