+++ lib/auth/services/index.dart
@@ -1,4 +1,4 @@
 export 'auth_services/auth_provider.dart';
 export 'auth_services/auth_service.dart';
 export 'auth_services/auth_user.dart';
-export 'firebase_auth_provider.dart';
+export 'auth_services/firebase_auth_provider.dart';
